moving back to array of primitive ints

+++ src/test/java/com/github/vy/fibertest/Util.java
@@ -14,13 +14,13 @@ class Util {
     static void testRingBenchmark(
             final int workerCount,
             final int ringSize,
-            final Integer[] sequences) {
+            final int[] sequences) {
         final int offset = workerCount - ringSize % workerCount;
         for (int i = 0; i < workerCount; i++)
             try {
                 assertEquals(
                         "sequence returned by Worker#" + i,
-                        -((offset + i) % workerCount), sequences[i].intValue());
+                        -((offset + i) % workerCount), sequences[i]);
             } catch (AssertionError ae) {
                 log.trace("sequences[] = {}", Arrays.toString(sequences));
                 throw ae;
